Updated 'README.md' to include a way to run EDITOR program.
Updated 'image_cropper.py' with minor changes to the code. Overall, it still works the same. Made the initial arguments to 'ImageCropper' to be set to 'None' by default.
Plan on adding to 'ImageCropper' method 'template_matching_multiple_methods()' to show a cropped image instead of just displaying the selected region of interest when template matching with different methods.

Added 'editor.py' file to be the main hub for the EDITOR program. This file mainly uses the 'ArgumentParser' object to look at the command-line arguments and decide what to do from there. At this time, the EDITOR can only crop images with the '-i' and '-ic' optional arguments.

Added 'string_functions.py' file to hold custom string functions I created.