Updated 'editor.py' with a new argument for the argument parser, allowing the EDITOR to edit files.
Added 'file_editor.py' to take care of editing files. Currently, it can only do a very specific thing of removing the first few characters of a line inside a document.

Updated 'string_functions.py' with a new function that gives choices to the user, asks for their input, and determine which choice they inputted.